moving level up logic to parent class, stealing from openings
