+++ src/assets/text/convos/modepostlevel0.csv
@@ -1,3 +1,3 @@
 0,SHOW_MONSTER,...,>,1
 1,SHOW_MONSTER|START_CHAT_MUSIC,I leveled up!,>,2
-4,SHOW_MONSTER,:),>,ModeMenu1
+2,SHOW_MONSTER,:),>,ModeMenu1
